Refactor common helper functions into util.go and minor cleanup
