+++ cmd/util.go
@@ -1,0 +1,83 @@
+// Copyright © 2017 UBC Launch Pad team@ubclaunchpad.com
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package cmd
+
+import (
+	"errors"
+	"os"
+	"path/filepath"
+	"strings"
+
+	git "gopkg.in/src-d/go-git.v4"
+)
+
+// checkForGit returns an error if we're not in a git repository.
+func checkForGit(cwd string) error {
+	// Quick failure if no .git folder.
+	gitFolder := filepath.Join(cwd, ".git")
+	if _, err := os.Stat(gitFolder); os.IsNotExist(err) {
+		return errors.New("this does not appear to be a git repository")
+	}
+
+	repo, err := git.PlainOpen(cwd)
+	if err != nil {
+		return err
+	}
+
+	remotes, err := repo.Remotes()
+
+	// Also fail if no remotes detected.
+	if len(remotes) == 0 {
+		return errors.New("there are no remotes associated with this repository")
+	}
+
+	return nil
+}
+
+// getLocalRepo gets the repo from disk.
+func getLocalRepo() (*git.Repository, error) {
+	cwd, err := os.Getwd()
+	if err != nil {
+		return nil, err
+	}
+	return git.PlainOpen(cwd)
+}
+
+// getSSHRemoteURL gets the URL of the given remote in the form
+// "git@github.com:[USER]/[REPOSITORY].git"
+func getSSHRemoteURL(r *git.Remote) string {
+	url := r.Config().URLs[0]
+	return strings.Replace(url, "https://github.com/", "git@github.com:", -1) + ".git"
+}
+
+// removeContents removes all files within given directory, returns nil if successful
+func removeContents(directory string) error {
+	d, err := os.Open(directory)
+	if err != nil {
+		return err
+	}
+	defer d.Close()
+	names, err := d.Readdirnames(-1)
+	if err != nil {
+		return err
+	}
+	for _, name := range names {
+		err = os.RemoveAll(filepath.Join(directory, name))
+		if err != nil {
+			return err
+		}
+	}
+	return nil
+}
